# internal/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Это REST API для задач.",
        "title": "Vado API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:5555",
    "basePath": "/",
    "paths": {
        "/tasks": {
            "get": {
                "description": "Возвращает все задачи",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tasks"
                ],
                "summary": "Получить список задач.",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.Task"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Добавляет новую задачу",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "tasks"
                ],
                "summary": "Создать задачу",
                "parameters": [
                    {
                        "description": "Task info",
                        "name": "task",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Task"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Task created.",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "description": "Чистит список задач",
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "tasks"
                ],
                "summary": "Удалить все задачи",
                "responses": {
                    "200": {
                        "description": "Delete tasks.",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/tasks/{id}": {
            "get": {
                "description": "Возвращает задачу по её ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tasks"
                ],
                "summary": "Получить задачу по ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID задачи",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Task"
                        }
                    },
                    "400": {
                        "description": "Invalid id",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Task not found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "description": "Удаляет задачу по её ID",
                "tags": [
                    "tasks"
                ],
                "summary": "Удалить задачу по ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID задачи",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Delete task {id}.",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Invalid id",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Task not found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "model.Task": {
            "description": "Структура задачи.",
            "type": "object",
            "properties": {
                "completed": {
                    "description": "Флаг выполнения",
                    "type": "boolean",
                    "example": false
                },
                "description": {
                    "description": "Описание задачи",
                    "type": "string",
                    "example": "Купить 2 литра"
                },
                "id": {
                    "description": "Уникальный идентификатор",
                    "type": "integer",
                    "format": "int64",
                    "example": 1
                },
                "name": {
                    "description": "Название задачи",
                    "type": "string",
                    "example": "Купить молоко"
                }
            }
        }
    }
}